Move `GetHeaderReps` and `mkHeaderRep` to `Internal.Header` so we can import it in other internal modules.
