+++ src/Servant/API/Routes/Header.hs
@@ -13,38 +13,4 @@ module Servant.API.Routes.Header
   )
 where
 
-import Data.Kind (Type)
-import Data.Typeable
-import GHC.TypeLits
-import Servant.API
 import "this" Servant.API.Routes.Internal.Header
-import "this" Servant.API.Routes.Utils
-
--- | Convenience function to construct a 'HeaderRep' from @sym :: 'Symbol'@ and @a :: Type'@.
-mkHeaderRep ::
-  forall sym a.
-  (KnownSymbol sym, Typeable a) =>
-  HeaderRep
-mkHeaderRep =
-  HeaderRep
-    { _hName = knownSymbolT @sym
-    , _hType = typeRepOf @a
-    }
-
-{- | Utility class to let us get a value-level list of 'HeaderRep's from a
-type-level list of 'Servant.API.Header.Header's. See the implementation of
-@'Servant.API.Route.HasRoutes' ('Verb' method status ctypes ('Headers' hs a))@ for an example.
--}
-class GetHeaderReps (hs :: [Type]) where
-  getHeaderReps :: [HeaderRep]
-
-instance GetHeaderReps '[] where
-  getHeaderReps = []
-
-instance
-  (GetHeaderReps rest, KnownSymbol h, Typeable v) =>
-  GetHeaderReps (Header h v ': rest)
-  where
-  getHeaderReps = header : getHeaderReps @rest
-    where
-      header = mkHeaderRep @h @v
